refactor: rename output column class to FinOpsColumn
Per review: FOCUSColumn is reserved for the FOCUS input format (consistent with CURColumn/AzureColumn). The columns SPRUCE generates for FinOps reporting become FinOpsColumn, and the azure.FOCUSColumns module becomes azure.FinOpsColumns. A Javadoc note records that some labels (RegionId, ChargeCategory) match the FOCUS input columns; sharing the label constants is left for when a FOCUS input format lands. No behaviour change.

Refs #238
