+++ src/main/java/com/digitalpebble/spruce/FinOpsColumn.java
@@ -1,0 +1,48 @@
+// SPDX-License-Identifier: Apache-2.0
+
+package com.digitalpebble.spruce;
+
+import org.apache.spark.sql.types.DataType;
+
+import java.util.Map;
+
+import static org.apache.spark.sql.types.DataTypes.DoubleType;
+import static org.apache.spark.sql.types.DataTypes.StringType;
+
+/**
+ * Provider-neutral columns generated by SPRUCE for FinOps reporting, named after the FOCUS
+ * (FinOps Open Cost &amp; Usage Specification) schema. The EnrichmentModules write these into the
+ * shared enriched values map, bridging the CSP-specific billing columns (cost, region, service,
+ * account, time) to neutral names used by the reporting scripts and dashboard.
+ *
+ * <p>Some labels (e.g. {@code RegionId}, {@code ChargeCategory}) hold the same values as those a
+ * FOCUS report carries on input. Once SPRUCE supports a FOCUS input format, those labels should be
+ * shared with the input column definition rather than duplicated here.
+ **/
+public class FinOpsColumn extends Column {
+
+    public static FinOpsColumn BILLED_COST = new FinOpsColumn("BilledCost", DoubleType);
+    public static FinOpsColumn BILLING_CURRENCY = new FinOpsColumn("BillingCurrency", StringType);
+    public static FinOpsColumn REGION_ID = new FinOpsColumn("RegionId", StringType);
+    public static FinOpsColumn SERVICE_NAME = new FinOpsColumn("ServiceName", StringType);
+    public static FinOpsColumn CHARGE_CATEGORY = new FinOpsColumn("ChargeCategory", StringType);
+    public static FinOpsColumn SUB_ACCOUNT_ID = new FinOpsColumn("SubAccountId", StringType);
+    public static FinOpsColumn CHARGE_PERIOD_START = new FinOpsColumn("ChargePeriodStart", StringType);
+    public static FinOpsColumn CHARGE_PERIOD_END = new FinOpsColumn("ChargePeriodEnd", StringType);
+    public static FinOpsColumn TAGS = new FinOpsColumn("Tags", StringType);
+
+    private FinOpsColumn(String l, DataType t) {
+        super(l, t);
+    }
+
+    /** Returns the String value for this column from the enriched values map, or null if absent. */
+    public String getString(Map<Column, Object> enrichedValues) {
+        return (String) enrichedValues.get(this);
+    }
+
+    /** Returns the Double value for this column from the enriched values map, or null if absent. */
+    public Double getDouble(Map<Column, Object> enrichedValues) {
+        return (Double) enrichedValues.get(this);
+    }
+
+}
