+++ src/test/java/com/digitalpebble/spruce/modules/azure/FinOpsColumnsTest.java
@@ -15,13 +15,13 @@ import org.junit.jupiter.api.Test;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.digitalpebble.spruce.FOCUSColumn.BILLED_COST;
-import static com.digitalpebble.spruce.FOCUSColumn.CHARGE_CATEGORY;
-import static com.digitalpebble.spruce.FOCUSColumn.CHARGE_PERIOD_END;
-import static com.digitalpebble.spruce.FOCUSColumn.CHARGE_PERIOD_START;
-import static com.digitalpebble.spruce.FOCUSColumn.REGION_ID;
-import static com.digitalpebble.spruce.FOCUSColumn.SERVICE_NAME;
-import static com.digitalpebble.spruce.FOCUSColumn.SUB_ACCOUNT_ID;
+import static com.digitalpebble.spruce.FinOpsColumn.BILLED_COST;
+import static com.digitalpebble.spruce.FinOpsColumn.CHARGE_CATEGORY;
+import static com.digitalpebble.spruce.FinOpsColumn.CHARGE_PERIOD_END;
+import static com.digitalpebble.spruce.FinOpsColumn.CHARGE_PERIOD_START;
+import static com.digitalpebble.spruce.FinOpsColumn.REGION_ID;
+import static com.digitalpebble.spruce.FinOpsColumn.SERVICE_NAME;
+import static com.digitalpebble.spruce.FinOpsColumn.SUB_ACCOUNT_ID;
 import static com.digitalpebble.spruce.SpruceColumn.REGION;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -31,9 +31,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * Maps Azure-native billing columns to provider-neutral FOCUS output columns.
  **/
-public class FOCUSColumnsTest {
+public class FinOpsColumnsTest {
 
-    private final FOCUSColumns module = new FOCUSColumns();
+    private final FinOpsColumns module = new FinOpsColumns();
     private final StructType schema = Utils.getSchema(module);
 
     @Test
@@ -130,9 +130,9 @@ public class FOCUSColumnsTest {
         assertTrue(java.util.Arrays.asList(module.columnsAdded()).contains(BILLED_COST));
         assertTrue(java.util.Arrays.asList(module.columnsAdded()).contains(CHARGE_PERIOD_END));
         assertFalse(java.util.Arrays.asList(module.columnsAdded())
-                .contains(com.digitalpebble.spruce.FOCUSColumn.TAGS));
+                .contains(com.digitalpebble.spruce.FinOpsColumn.TAGS));
         assertFalse(java.util.Arrays.asList(module.columnsAdded())
-                .contains(com.digitalpebble.spruce.FOCUSColumn.BILLING_CURRENCY));
+                .contains(com.digitalpebble.spruce.FinOpsColumn.BILLING_CURRENCY));
     }
 
     /**
